+++ src/chanjo2/models/sql_models.py
@@ -23,7 +23,7 @@ class Sample(Base):
     __tablename__ = "samples"
 
     id = Column(Integer, primary_key=True, index=True)
-    name = Column(String(64), nullable=False, unique=True)
+    name = Column(String(64), nullable=False, unique=True, index=True)
     display_name = Column(String(64), nullable=True, unique=False)
     case_id = Column(Integer, ForeignKey("cases.id"), nullable=False)
     coverage_file_path = Column(String, nullable=False)