+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Colored logs for development and debugging
 - Common test fixtures
 - Some badges on README page
+- Tests for the endpoints
 ### Fixed
 - Bugs preventing the gunicorn app to launch
 - Code to compose DB url to work when app is invoked from docker-compose